+++ src/routes/index.tsx
@@ -8,7 +8,7 @@ import Profile from '@/routes/pages/Profile'
 import PlaylistDetail from '@/routes/pages/PlaylistDetail'
 import SavedPlaylistDetail from '@/routes/pages/SavedPlaylistsDetail'
 import RequiresAuth from '@/routes/protected/RequiresAuth'
-import EditPlaylist from '@/routes/pages/EditPlaylist'
+import EditPlaylist from '@/components/EditPlaylistModal'
 import SignIn from './pages/SignIn'
 import NotFound from './pages/NotFound'
 import dayjs from 'dayjs'
